app/requirements.txt
asgiref==3.7.2
Django==4.2.2
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
psycopg2-binary==2.9.6
python-decouple==3.8
# me lo pide el campo de ImageField
Pillow==10.0.0 
djangorestframework==3.14.0
django-filter==23.2
Markdown==3.4.4